Create New Experiment

Creates an experiment object which will contain all data and metadata. This object will be saved and timestamped by date, so if working with this file over different dates, the object will not be overwritten as a new object will be created (to preserve backups).

create_experiment(experiment_name)

Arguments

experiment_name

The name of the experiment

Value

A list object containing statistics and metadata for the entire experiment. The `groups` and `experimenters` field will auto-populate based on added data using the `update_experiment()` function.

Examples

create_experiment(experiment_name = "My-Project")
#> Creating new experiment...
#> $name
#> [1] "My-Project"
#> 
#> $last_saved
#> [1] "2022-06-16 14:34:46 UTC"
#> 
#> $groups
#> NULL
#> 
#> $experimenters
#> NULL
#> 
#> $experimental_data
#> list()
#>